[Project] Prefactor: extract defaulting into dedicated file
